UI: factored out FormInput
