+++ src/UI/Component/Input/Field/Input.php
@@ -28,81 +28,8 @@ use ILIAS\UI\Component\OnUpdateable;
  * into other types of data. This means, that e.g. the value of an input could
  * be some id, while the content could be some object referenced by that id.
  */
-interface Input extends Component, JavaScriptBindable, OnUpdateable
+interface Input
 {
-
-    /**
-     * Get the label of the input.
-     *
-     * @return    string
-     */
-    public function getLabel();
-
-
-    /**
-     * Get an input like this, but with a replaced label.
-     *
-     * @param    string $label
-     *
-     * @return    Input
-     */
-    public function withLabel($label);
-
-
-    /**
-     * Get the byline of the input.
-     *
-     * @return    string|null
-     */
-    public function getByline();
-
-
-    /**
-     * Get an input like this, but with an additional/replaced label.
-     *
-     * @param    string|null $byline
-     *
-     * @return    Input
-     */
-    public function withByline($byline);
-
-
-    /**
-     * Is this field required?
-     *
-     * @return    bool
-     */
-    public function isRequired();
-
-
-    /**
-     * Get an input like this, but set the field to be required (or not).
-     *
-     * @param    bool $is_required
-     *
-     * @return    Input
-     */
-    public function withRequired($is_required);
-
-
-    /**
-     * Is this input disabled?
-     *
-     * @return    bool
-     */
-    public function isDisabled();
-
-
-    /**
-     * Get an input like this, but set it to a disabled state.
-     *
-     * @param    bool $is_disabled
-     *
-     * @return    Input
-     */
-    public function withDisabled($is_disabled);
-
-
     /**
      * Get the value that is displayed in the input client side.
      *
@@ -124,24 +51,6 @@ interface Input extends Component, JavaScriptBindable, OnUpdateable
 
 
     /**
-     * The error of the input as used in HTML.
-     *
-     * @return string|null
-     */
-    public function getError();
-
-
-    /**
-     * Get an input like this one, with a different error.
-     *
-     * @param    string
-     *
-     * @return    Input
-     */
-    public function withError($error);
-
-
-    /**
      * Apply a transformation to the content of the input.
      *
      * @param    Transformation $trafo
@@ -149,19 +58,4 @@ interface Input extends Component, JavaScriptBindable, OnUpdateable
      * @return    Input
      */
     public function withAdditionalTransformation(Transformation $trafo);
-
-
-    /**
-     * Get update code
-     *
-     * This method has to return JS code that calls
-     * il.UI.filter.onFieldUpdate(event, '$id', string_value);
-     * - initially "onload" and
-     * - on every input change.
-     * It must pass a readable string representation of its value in parameter 'string_value'.
-     *
-     * @param \Closure $binder
-     * @return string
-     */
-    public function getUpdateOnLoadCode() : \Closure;
 }
